+++ Modelos/Repartidores.py
@@ -7,7 +7,7 @@ from Modelos.Database.Restaurant import Restaurant
 Base = declarative_base()
 class Estados(Base):
     
-    __tablename__ = 'Administradores'
+    __tablename__ = 'Repartidores'
     
     id_repartidor = Column(Integer(), primary_key=True, autoincrement=True)
     nombre_repartidor = Column(String(50), nullable = False)
@@ -16,3 +16,5 @@ class Estados(Base):
     
     def __str__(self):
         return  "ID: " + self.id_repartidor 
+
+Base.metadata.create_all(Restaurant.getInstance().getEngine())
